requirements.txt
httpx
click
PyYAML
Jinja2
